Task1 Upload
Task1 consists of a fully fuctional website of the disaster alleviation foundation, where it enables its users to send donations according to 2 categories (Monetary and Goods), this is only accessable after signing up and logging in. The website also allowed users to caputure current disasters by giving dates, location and required aids. There is also the admin side of the website, where the admin will have to login with their credentails, to access the list of all donations and disasters, as well as edit or delete them. Together with this, there is a home page, which shows and overview of the organization and various images can be found on this page.